+++ pyscf/dh/udfdh.py
@@ -2,7 +2,8 @@
 from pyscf.dh.dh import DHBase
 from pyscf.dh.dhutil import gen_batch, calc_batch_size, timing, tot_size, hermi_sum_last2dim, xc_equal
 from pyscf.dh.rdfdh import kernel, energy_nuc, energy_tot
-from pyscf.dh.mp2_ajz import get_cderi_mo, energy_elec_ump2_ajz, energy_elec_mp2_dfump2
+from pyscf.dh.mp2_ajz import get_cderi_mo, energy_elec_ump2_ajz
+from pyscf.dh.dh import energy_elec_mp2_dfump2_native, energy_elec_mp2_dfump2
 # pyscf import
 from pyscf import lib, gto, df, dft, scf
 from pyscf.scf import ucphf
@@ -46,6 +47,9 @@ def energy_elec_pt2(mf, params=None, eng_bi=None, **kwargs):
     eng_bi1, eng_bi2 = eng_bi if eng_bi else mf.energy_elec_mp2(eval_ss=mf.eval_ss, **kwargs)
     if getattr(mf, 'mp2_backend', None) == "dfmp2_native":
         return cc * eng_bi1, None, None
+    if getattr(mf, 'mp2_backend', None) == "dfmp2":
+        e_corr_os, e_corr_ss = eng_bi1, eng_bi2
+        return cc * (c_os * e_corr_os + c_ss * e_corr_ss), e_corr_os, e_corr_ss
     eng_os = eng_bi1[αβ]
     eng_ss = 0.5 * (eng_bi1[αα] + eng_bi1[ββ] - eng_bi2[αα] - eng_bi2[ββ])
     eng_pt2 = cc * (c_os * eng_os + c_ss * eng_ss)
@@ -182,8 +186,9 @@ class UDFDH(DHBase):
                  grids: dft.Grids = None,
                  grids_cpks: dft.Grids = None,
                  mp2_backend: str = "ajz",
+                 frozen: int = None,
                  ):
-        super().__init__(mol, xc, auxbasis_jk, auxbasis_ri, mp2_backend)
+        super().__init__(mol, xc, auxbasis_jk, auxbasis_ri, mp2_backend, frozen)
         mf_s = dft.UKS(mol, xc=self.xc).density_fit(auxbasis=self.auxbasis_jk)
         self.grids = grids if grids else mf_s.grids
         self.grids_cpks = grids_cpks if grids_cpks else self.grids
@@ -208,6 +213,8 @@ class UDFDH(DHBase):
         self.nmo = self.nao
         self.nvir = (self.nmo - self.nocc[α], self.nmo - self.nocc[β])
         if mp2_backend == "dfmp2_native":
+            self.energy_elec_mp2 = partial(energy_elec_mp2_dfump2_native, self)
+        elif mp2_backend == "dfmp2":
             self.energy_elec_mp2 = partial(energy_elec_mp2_dfump2, self)
         else:
             self.energy_elec_mp2 = partial(energy_elec_ump2_ajz, self)
